requirements.txt
openai
requests
python-dotenv
elevenlabs
